+++ src/renderer-lyric/App.vue
@@ -13,14 +13,14 @@
       </transition>
     </div>
     <template v-if="isShowResize">
-      <div class="resize resize-left" @mousedown.self="handleMouseDown('left', $event)" @touchstart.self="handleTouchDown('left', $event)" />
-      <div class="resize resize-top" @mousedown.self="handleMouseDown('top', $event)" @touchstart.self="handleTouchDown('top', $event)" />
-      <div class="resize resize-right" @mousedown.self="handleMouseDown('right', $event)" @touchstart.self="handleTouchDown('right', $event)" />
-      <div class="resize resize-bottom" @mousedown.self="handleMouseDown('bottom', $event)" @touchstart.self="handleTouchDown('bottom', $event)" />
-      <div class="resize resize-top-left" @mousedown.self="handleMouseDown('top-left', $event)" @touchstart.self="handleTouchDown('top-left', $event)" />
-      <div class="resize resize-top-right" @mousedown.self="handleMouseDown('top-right', $event)" @touchstart.self="handleTouchDown('top-right', $event)" />
-      <div class="resize resize-bottom-left" @mousedown.self="handleMouseDown('bottom-left', $event)" @touchstart.self="handleTouchDown('bottom-left', $event)" />
-      <div class="resize resize-bottom-right" @mousedown.self="handleMouseDown('bottom-right', $event)" @touchstart.self="handleTouchDown('bottom-right', $event)" />
+      <div class="resize resize-left" @mousedown.self="handleMouseDown('left', $event)" @touchstart.passive.self="handleTouchDown('left', $event)" />
+      <div class="resize resize-top" @mousedown.self="handleMouseDown('top', $event)" @touchstart.passive.self="handleTouchDown('top', $event)" />
+      <div class="resize resize-right" @mousedown.self="handleMouseDown('right', $event)" @touchstart.passive.self="handleTouchDown('right', $event)" />
+      <div class="resize resize-bottom" @mousedown.self="handleMouseDown('bottom', $event)" @touchstart.passive.self="handleTouchDown('bottom', $event)" />
+      <div class="resize resize-top-left" @mousedown.self="handleMouseDown('top-left', $event)" @touchstart.passive.self="handleTouchDown('top-left', $event)" />
+      <div class="resize resize-top-right" @mousedown.self="handleMouseDown('top-right', $event)" @touchstart.passive.self="handleTouchDown('top-right', $event)" />
+      <div class="resize resize-bottom-left" @mousedown.self="handleMouseDown('bottom-left', $event)" @touchstart.passive.self="handleTouchDown('bottom-left', $event)" />
+      <div class="resize resize-bottom-right" @mousedown.self="handleMouseDown('bottom-right', $event)" @touchstart.passive.self="handleTouchDown('bottom-right', $event)" />
     </template>
     <layout-icons />
   </div>