+++ src/renderer-lyric/components/layout/LyricHorizontal/useLyric.js
@@ -210,7 +210,7 @@ export default (isComputeHeight) => {
   onMounted(() => {
     document.addEventListener('mousemove', handleMouseMsMove)
     document.addEventListener('mouseup', handleMouseMsUp)
-    document.addEventListener('touchmove', handleTouchMove)
+    document.addEventListener('touchmove', handleTouchMove, { passive: true })
     document.addEventListener('touchend', handleMouseMsUp)
 
     initLrc(lyric.lines, null)